feat(ui-react): wire up command palette and refactor onto base dialog
The Cmd/Ctrl+K command palette was orphaned — nothing rendered it, so the
shortcut did nothing. Mount it once at the AppLayout root so it works
app-wide, and rebuild it on the shared BaseDialog primitive to get a real
focus trap, native Escape handling, and a centered top-layer modal instead
of the hand-rolled overlay.

Correctness fixes:
- guard arrow-key navigation against an empty result list (was % 0 = NaN)
- derive the active index during render instead of writing state in render
- match the shortcut case-insensitively so Shift/CapsLock still toggles

Accessibility: adopt the WAI-ARIA combobox/listbox pattern — the input is a
combobox driving aria-activedescendant, options are non-tabbable listbox
rows, the empty state is a polite live region, and decorative icons/kbd are
hidden from assistive tech. Tighten useFocusTrap to exclude tabindex="-1"
so the trap never lands on the activedescendant-driven options.
