+++ ui/apps/console/src/components/layout/AppLayout.tsx
@@ -9,6 +9,7 @@ import AnnouncementModalTrigger from "../announcements/AnnouncementModalTrigger"
 import DeviceChooserTrigger from "../billing/DeviceChooserTrigger";
 import { SidebarMobileDrawer } from "./SidebarShell";
 import ChatwootProvider from "./ChatwootProvider";
+import CommandPalette from "@/components/commandPalette/CommandPalette";
 import { useNamespaces } from "@/hooks/useNamespaces";
 import { useTerminalStore } from "@/stores/terminalStore";
 import { useSidebarLayout } from "@/hooks/useSidebarLayout";
@@ -80,6 +81,7 @@ export default function AppLayout() {
           </div>
         </div>
         <TerminalManager sidebarOffset={sidebarOffset} />
+        <CommandPalette />
         <WelcomeWizardTrigger />
         <AnnouncementModalTrigger />
         <DeviceChooserTrigger />
