+++ ui/apps/console/src/hooks/useFocusTrap.ts
@@ -1,7 +1,10 @@
 import { RefObject, useEffect } from "react";
 
-const FOCUSABLE
-  = 'a[href],button:not([disabled]),input:not([disabled]),select:not([disabled]),textarea:not([disabled]),[tabindex]:not([tabindex="-1"])';
+// Every clause excludes tabindex="-1": an element explicitly removed from the
+// tab order must never receive focus from the trap (e.g. listbox options that
+// are driven by aria-activedescendant rather than roving DOM focus).
+const FOCUSABLE =
+  ':is(a[href],button:not([disabled]),input:not([disabled]),select:not([disabled]),textarea:not([disabled]),[tabindex]):not([tabindex="-1"])';
 
 /**
  * Traps keyboard focus within `containerRef` while `active` is true.
